requirements.txt
openai>=1.0.0
python-dotenv
rich
pydantic>=2.0.0
pytest